massive code re-organization
