+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/LocateClassCommand.java
@@ -1,7 +1,7 @@
 package edu.ucsc.refactor.cli.commands;
 
-import edu.ucsc.refactor.internal.ClassUnit;
-import edu.ucsc.refactor.spi.ProgramUnit;
+import edu.ucsc.refactor.locators.ClassUnit;
+import edu.ucsc.refactor.ProgramUnit;
 import io.airlift.airline.Command;
 
 /**
