+++ ext/vesper-cli/src/edu/ucsc/refactor/cli/commands/LocateParamCommand.java
@@ -1,7 +1,7 @@
 package edu.ucsc.refactor.cli.commands;
 
-import edu.ucsc.refactor.internal.ParameterUnit;
-import edu.ucsc.refactor.spi.ProgramUnit;
+import edu.ucsc.refactor.locators.ParameterUnit;
+import edu.ucsc.refactor.ProgramUnit;
 import io.airlift.airline.Command;
 
 /**
