+++ ext/vesper-history/src/edu/ucsc/refactor/internal/NavigableJavaRefactorer.java
@@ -5,7 +5,6 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import edu.ucsc.refactor.*;
-import edu.ucsc.refactor.spi.UnitLocator;
 import edu.ucsc.refactor.util.Commit;
 import edu.ucsc.refactor.util.CommitHistory;
 import edu.ucsc.refactor.util.CommitPublisher;
@@ -76,7 +75,7 @@ public class NavigableJavaRefactorer implements NavigableRefactorer {
 
 
     @Override public Set<Issue> detectIssues(Source code) {
-        final Introspector introspector = refactorer.getIntrospector(code);
+        final Introspector introspector = Vesper.createIntrospector();
         try {
             final Set<Issue> issues = introspector.detectIssues(code);
             for(Issue each : issues){
@@ -103,14 +102,6 @@ public class NavigableJavaRefactorer implements NavigableRefactorer {
     }
 
 
-    @Override public Introspector getIntrospector() {
-        return refactorer.getIntrospector();
-    }
-
-    @Override public Introspector getIntrospector(Source src) {
-        return refactorer.getIntrospector(src);
-    }
-
     @Override public List<Issue> getIssues(Source key) {
         if(getIssueRegistry().containsKey(key)){
             return getIssueRegistry().get(key);
@@ -136,10 +127,6 @@ public class NavigableJavaRefactorer implements NavigableRefactorer {
         );
     }
 
-
-    @Override public UnitLocator getLocator(Source src) {
-        return refactorer.getLocator(src);
-    }
 
     @Override public boolean hasIssues(Source code) {
         return !getIssues(Preconditions.checkNotNull(code)).isEmpty();
