+++ src/edu/ucsc/refactor/ProgramUnit.java
@@ -1,7 +1,4 @@
-package edu.ucsc.refactor.spi;
-
-import edu.ucsc.refactor.Context;
-import edu.ucsc.refactor.NamedLocation;
+package edu.ucsc.refactor;
 
 import java.util.List;
