+++ src/edu/ucsc/refactor/locators/MethodUnit.java
@@ -1,9 +1,10 @@
-package edu.ucsc.refactor.internal;
+package edu.ucsc.refactor.locators;
 
 import com.google.common.base.Preconditions;
 import edu.ucsc.refactor.Context;
 import edu.ucsc.refactor.Location;
 import edu.ucsc.refactor.NamedLocation;
+import edu.ucsc.refactor.internal.ProgramUnitLocation;
 import edu.ucsc.refactor.internal.util.AstUtil;
 import org.eclipse.jdt.core.dom.ASTNode;
 import org.eclipse.jdt.core.dom.MethodDeclaration;
