+++ src/edu/ucsc/refactor/spi/graph/GraphUtils.java
@@ -1,4 +1,4 @@
-package edu.ucsc.refactor.util.graph;
+package edu.ucsc.refactor.spi.graph;
 
 import java.util.ArrayList;
 import java.util.List;
