+++ test/edu/ucsc/refactor/RefactorersTest.java
@@ -1,7 +1,6 @@
 package edu.ucsc.refactor;
 
-import edu.ucsc.refactor.internal.*;
-import edu.ucsc.refactor.spi.UnitLocator;
+import edu.ucsc.refactor.locators.*;
 import edu.ucsc.refactor.util.Recommender;
 import org.junit.Test;
 
@@ -44,8 +43,7 @@ public class RefactorersTest {
     }
 
     @Test public void testRefactorerIssueDetection(){
-        final Refactorer refactorer = Vesper.createRefactorer();
-        final Introspector introspector = refactorer.getIntrospector(SRC);
+        final Introspector introspector = Vesper.createIntrospector();
         final Set<Issue> issues = introspector.detectIssues(SRC);
         assertThat(issues.isEmpty(), is(false));
     }
@@ -53,7 +51,7 @@ public class RefactorersTest {
     @Test public void testRefactorerCreateChanges() {
         // for issue and for edit
         final Refactorer refactorer = Vesper.createRefactorer();
-        final Introspector introspector = refactorer.getIntrospector(SRC);
+        final Introspector introspector = Vesper.createIntrospector();
         final Set<Issue> issues     = introspector.detectIssues(SRC);
 
         assertThat(issues.isEmpty(), is(false));
@@ -85,7 +83,7 @@ public class RefactorersTest {
     @Test public void testRefactorerRecommendChange(){
         // for any issues found on the SRC
         final Refactorer refactorer = Vesper.createRefactorer();
-        final Introspector introspector = refactorer.getIntrospector(SRC);
+        final Introspector introspector = Vesper.createIntrospector();
         final List<Change> changes = Recommender.recommendChanges(refactorer, SRC, introspector.detectIssues(SRC));
 
         assertThat(changes.isEmpty(), is(false));
@@ -93,8 +91,7 @@ public class RefactorersTest {
     }
 
     @Test public void testRefactorerUnitLocator() {
-        final Refactorer    refactorer  = Vesper.createRefactorer();
-        final UnitLocator   locator     = refactorer.getLocator(SRC);
+        final UnitLocator   locator = Vesper.createUnitLocator(SRC);
 
         final List<NamedLocation> params = locator.locate(new ParameterUnit("message"));
         assertThat(params.isEmpty(), is(false));
